bugs fixed, ready

+++ components/renderTaskList.js
@@ -1,7 +1,6 @@
-import { listOfAchives } from "../utils/archive.js";
+import { listOfAchives } from "../utils/store.js";
 import onClickArchive from "../handlers/onClickArchive.js";
 import onClickDelete from "../handlers/onClickDelete.js";
-import onClickEdit from "../handlers/onCLickEdit.js";
 import { SecondTable } from "./SecondTable.js";
 import { onClickZoomContent } from "../handlers/onClickZoomContent.js";
 
@@ -11,27 +10,36 @@ export const categories = {
   Gym: "gym.svg",
   Idea: "idea.svg",
 };
+let isArchive = false;
 
 export const renderList = () => {
   const mainItems = document.querySelector(".list__items__main ");
+  
   mainItems.innerHTML = "";
+
   listOfAchives.forEach((key, i) => {
-    if (!key.archived)
-      mainItems.innerHTML += `<ul class="taskContainer">
+    //render archived tasks or unarchived
+    if (!isArchive) {
+      if (!key.archived)
+        mainItems.innerHTML += `<ul class="taskContainer">
       <li class ="task"><img class="taskImg" src='./assets/img/${
         categories[key.category]
       }' alt='${key.category}'/>${key.name}</li>
       <li class ="task">${key.created}</li>
       <li class ="task">${key.category}</li>
-      <li name="content" id='${key.id}' class="task contentOfTask" data-modal="#modal_2">${
-        key.content.length > 30 ? key.content.slice(0, 30) + "..." : key.content
-      }</li>
-      <li class ="task">${key.date}</li>
+      <li name="content" id='${
+        key.id
+      }' class="task contentOfTask editZoom" data-modal="#modal_2">${
+          key.content.length > 30
+            ? key.content.slice(0, 30) + "..."
+            : key.content
+        }</li>
+      <li class ="task dateTask">${key.date}</li>
       <li class="sidebar">
       <img class='sidebar__buttons' src='./assets/img/archive.svg' name="archiveItem" id='${
         key.id
       }' alt="archive">
-      <img class='sidebar__buttons' src= "./assets/img/pencil.svg" name="editItem" id='${
+      <img class='sidebar__buttons editZoom'  data-modal="#modal_2" src= "./assets/img/pencil.svg" name="editItem" id='${
         key.id
       }' alt="pencil">
       <img class='sidebar__buttons' src= "./assets/img/trash.svg" name="deleteItem" id='${
@@ -39,10 +47,48 @@ export const renderList = () => {
       }' alt="trash">
       </li>
       </ul>`;
+    } else {
+      if (key.archived)
+        mainItems.innerHTML += `<ul class="taskContainer">
+        <li class ="task"><img class="taskImg" src='./assets/img/${
+          categories[key.category]
+        }' alt='${key.category}'/>${key.name}</li>
+        <li class ="task">${key.created}</li>
+        <li class ="task">${key.category}</li>
+        <li name="content" id='${
+          key.id
+        }' class="task contentOfTask editZoom" data-modal="#modal_2">${
+          key.content.length > 30
+            ? key.content.slice(0, 30) + "..."
+            : key.content
+        }</li>
+        <li class ="task dateTask">${key.date}</li>
+        <li class="sidebar">
+        <img class='sidebar__buttons' src='./assets/img/unarchive.svg' name="unarchiveItem" id='${
+          key.id
+        }' alt="archive">
+        <img class='sidebar__buttons editZoom'  data-modal="#modal_2" src= "./assets/img/pencil.svg" name="editItem" id='${
+          key.id
+        }' alt="pencil">
+        <img class='sidebar__buttons' src= "./assets/img/trash.svg" name="deleteItem" id='${
+          key.id
+        }' alt="trash">
+        </li>
+        </ul>`;
+    }
   });
-  console.log("renderList");
   const archiveItemButton = document.querySelectorAll('[name="archiveItem"]');
   archiveItemButton.forEach((key, i) => {
+    key.addEventListener("click", () => {
+      onClickArchive(+key.id);
+      renderList();
+      SecondTable();
+    });
+  });
+  const unArchiveItemButton = document.querySelectorAll(
+    '[name="unarchiveItem"]'
+  );
+  unArchiveItemButton.forEach((key, i) => {
     key.addEventListener("click", () => {
       onClickArchive(+key.id);
       renderList();
@@ -61,7 +107,7 @@ export const renderList = () => {
   const editItemButton = document.querySelectorAll('[name="editItem"]');
   editItemButton.forEach((key, i) => {
     key.addEventListener("click", () => {
-      onClickEdit(+key.id);
+      onClickZoomContent(+key.id);
       renderList();
       SecondTable();
     });
@@ -69,7 +115,13 @@ export const renderList = () => {
   const content = document.querySelectorAll('[name="content"]');
   content.forEach((key) => {
     key.addEventListener("click", () => {
-      onClickZoomContent(+key.id)
+      onClickZoomContent(+key.id);
     });
   });
 };
+const showArchiveBtn = document.querySelector(".sidebar__buttons__archive");
+showArchiveBtn.addEventListener("click", () => {
+  showArchiveBtn.innerHTML = isArchive ? "Active" : "Archived";
+  isArchive = !isArchive;
+  renderList();
+});
